# [opt](nereids) optimize small sql