--- a/fe/fe-core/src/main/java/org/apache/doris/catalog/FunctionRegistry.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/catalog/FunctionRegistry.java
@@ -377,7 +377,6 @@ public class FunctionRegistry {
         @Override
         public Expression withChildren(List<Expression> children) {
             throw new AnalysisException("could not call withChildren on UdfSignatureSearcher");
-
         }
     }
 }