--- a/fe/fe-core/src/main/java/org/apache/doris/catalog/TableIf.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/catalog/TableIf.java
@@ -236,10 +236,14 @@ public interface TableIf {
 
     default Set<PrimaryKeyConstraint> getPrimaryKeyConstraints() {
         try {
-            return getConstraintsMapUnsafe().values().stream()
-                    .filter(PrimaryKeyConstraint.class::isInstance)
-                    .map(PrimaryKeyConstraint.class::cast)
-                    .collect(ImmutableSet.toImmutableSet());
+            ImmutableSet.Builder<PrimaryKeyConstraint> constraintBuilder = ImmutableSet.builder();
+            for (Constraint constraint : getConstraintsMapUnsafe().values()) {
+                if (!(constraint instanceof PrimaryKeyConstraint)) {
+                    continue;
+                }
+                constraintBuilder.add((PrimaryKeyConstraint) constraint);
+            }
+            return constraintBuilder.build();
         } catch (Exception ignored) {
             return ImmutableSet.of();
         }
@@ -247,10 +251,14 @@ public interface TableIf {
 
     default Set<UniqueConstraint> getUniqueConstraints() {
         try {
-            return getConstraintsMapUnsafe().values().stream()
-                    .filter(UniqueConstraint.class::isInstance)
-                    .map(UniqueConstraint.class::cast)
-                    .collect(ImmutableSet.toImmutableSet());
+            ImmutableSet.Builder<UniqueConstraint> constraintBuilder = ImmutableSet.builder();
+            for (Constraint constraint : getConstraintsMapUnsafe().values()) {
+                if (!(constraint instanceof UniqueConstraint)) {
+                    continue;
+                }
+                constraintBuilder.add((UniqueConstraint) constraint);
+            }
+            return constraintBuilder.build();
         } catch (Exception ignored) {
             return ImmutableSet.of();
         }
@@ -493,9 +501,9 @@ public interface TableIf {
     }
 
     default String getNameWithFullQualifiers() {
-        return String.format("%s.%s.%s", getDatabase().getCatalog().getName(),
-                ClusterNamespace.getNameFromFullName(getDatabase().getFullName()),
-                getName());
+        return getDatabase().getCatalog().getName()
+                + "." + ClusterNamespace.getNameFromFullName(getDatabase().getFullName())
+                + "." + getName();
     }
 
     default boolean isManagedTable() {