--- a/fe/fe-core/src/main/java/org/apache/doris/nereids/CascadesContext.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/nereids/CascadesContext.java
@@ -542,6 +542,10 @@ public class CascadesContext implements ScheduleContext {
         return isEnableExprTrace;
     }
 
+    public boolean rewritePlanContainsTypes(Class<?>... types) {
+        return getRewritePlan().containsType(types);
+    }
+
     public RuntimeFilterContextV2 getRuntimeFilterV2Context() {
         return runtimeFilterV2Context;
     }