--- a/fe/fe-core/src/main/java/org/apache/doris/nereids/StatementContext.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/nereids/StatementContext.java
@@ -452,6 +452,10 @@ public class StatementContext implements Closeable {
         return exprIdGenerator.getNextId();
     }
 
+    public IdGenerator<ExprId> getExprIdGenerator() {
+        return exprIdGenerator;
+    }
+
     public CTEId getNextCTEId() {
         return cteIdGenerator.getNextId();
     }