--- a/fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/algebra/CatalogRelation.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/algebra/CatalogRelation.java
@@ -43,5 +43,4 @@ public interface CatalogRelation extends Relation {
     default List<Slot> getOperativeSlots() {
         return ImmutableList.of();
     }
-
 }