… fe/fe-core/src/main/java/org/apache/doris/nereids/glue/translator/PhysicalPlanTranslator.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/executor/AbstractBatchJobExecutor.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/executor/Analyzer.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/executor/Rewriter.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/AdaptiveTopDownRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/BottomUpVisitorRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/CostBasedRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/CustomRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/PlanTreeRewriteBottomUpJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/PlanTreeRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/PlanTreeRewriteTopDownJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/ProcessState.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/RewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/RewriteTopDownJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/RootPlanTreeRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/TopDownVisitorRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/TopicRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/scheduler/SimpleJobScheduler.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/load/NereidsLoadScanProvider.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/load/NereidsLoadUtils.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/memo/Group.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/memo/GroupExpression.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/memo/Memo.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/memo/StructInfoMap.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/minidump/MinidumpUtils.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/parser/LogicalPlanBuilder.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/pattern/ExpressionPatternRules.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/pattern/ExpressionPatternTraverseListeners.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/pattern/PatternDescriptor.java
… fe/fe-core/src/main/java/org/apache/doris/nereids/pattern/PatternMatcher.java