### What problem does this PR solve?

Problem Summary:

this pr improve QPS for small sql, by this points:
1. try best use **Visitor** rewriter to traverse plan, instead of **Stack** rewriter
2. fast skip traverse plan tree, if the tree not contains the specify type. for example, if not contains LogicalApply, we will not traverse for the rules of `Subquery unnesting` topic 
3. use **RoaringBitmap** to check whether the slot is exist, instead of `Set<Slot>`
4. skip prune column for logicalFilter(logicalOlapScan), or else need more rules to optimize it. for example, logicalProject(logicalFilter(localcalOlapScan)) → ColumnPruning →  logicalProject(logicalFilter(logicalProject(localcalOlapScan))) → PushDownFilterThroughProject → logicalProject(logicalProject(logicalFilter(localcalOlapScan))) → MergeProjects → logicalProject(logicalFilter(localcalOlapScan)))
5. use for loop instead of collection's stream api
6. faster process string, like format, split
7.  try to process project in one rule as possible, for example, project(project(oneRowRelatoin)) → oneRowRelation
8. fast compute some hashCode, e.g. LogicalOlapScan only use relationId as hash code
9. reuse ConnectContext to generate next exprId without invoke ConnectContext.get() multiple times
10. don't find table multiple times for compute lots of ColumnStatistic which have same table


this sql can optimize from 21000 QPS to 32000 QPS
```sql
explain
select date
from test.tbl
```

this sql can optimize from 16000 QPS to 23000 QPS
```sql
explain
select col1, col2, ..., col16
from test.tbl2
where col1 = 2 and col2 = 'xxx'
  and col3 >= 1721480971 and col3 <= 1721507345
  and col4>=20240720 and col4<=20240721 ;
```

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [ ] Manual test (add detailed scripts or steps below)
    - [ ] No need to test or manual test. Explain why:
        - [x] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [x] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [x] No.
    - [ ] Yes. <!-- Add document PR link here. eg: https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR should merge into -->