@hello-stephen 2025-04-22
# BE Regression P0 && UT Coverage Report
Increment line coverage `0.00% (0/5)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_269d649f492be5f878643cc388156f5cd46c3a29_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_269d649f492be5f878643cc388156f5cd46c3a29_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage |                    |
| Line Coverage     |                    |
| Region Coverage   |                    |
| Branch Coverage   |                    |