@hello-stephen 2025-04-22
# BE UT Coverage Report
Increment line coverage `45.45% (5/11)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/e433e996df7f2faa30b47bac7e57ed0f33338466_e433e996df7f2faa30b47bac7e57ed0f33338466/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/e433e996df7f2faa30b47bac7e57ed0f33338466_e433e996df7f2faa30b47bac7e57ed0f33338466/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 53.54% (14536/27148) |
| Line Coverage     | 42.39% (126042/297364) |
| Region Coverage   | 41.18% (64396/156358) |
| Branch Coverage   | 35.77% (32375/90514) |