@hello-stephen 2025-05-30
# BE Regression && UT Coverage Report
Increment line coverage `100.00% (3/3)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_4f545e1155c5213c9b291cefcb9270d3c285fe3f_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_4f545e1155c5213c9b291cefcb9270d3c285fe3f_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 59.12% (15602/26392) |
| Line Coverage     | 48.74% (145456/298434) |
| Region Coverage   | 46.06% (82948/180089) |
| Branch Coverage   | 39.88% (40819/102342) |