@hello-stephen 2025-06-03
# BE Regression && UT Coverage Report
Increment line coverage `100.00% (3/3)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_57c0929461f5112a8c32541042f197049663c8c9_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_57c0929461f5112a8c32541042f197049663c8c9_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 77.37% (20412/26383) |
| Line Coverage     | 69.69% (208049/298517) |
| Region Coverage   | 67.88% (122275/180126) |
| Branch Coverage   | 61.22% (62675/102372) |