@hello-stephen 2025-06-03
# BE UT Coverage Report
Increment line coverage `100.00% (3/3)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/891ffdd0999a355bf0be8ad2679c0a65a9394cac_891ffdd0999a355bf0be8ad2679c0a65a9394cac/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/891ffdd0999a355bf0be8ad2679c0a65a9394cac_891ffdd0999a355bf0be8ad2679c0a65a9394cac/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 56.18% (15052/26791) |
| Line Coverage     | 44.98% (134318/298595) |
| Region Coverage   | 44.09% (67571/153246) |
| Branch Coverage   | 38.64% (34620/89590) |