@hello-stephen 2025-06-03
# BE Regression && UT Coverage Report
Increment line coverage `100.00% (3/3)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_891ffdd0999a355bf0be8ad2679c0a65a9394cac_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_891ffdd0999a355bf0be8ad2679c0a65a9394cac_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 79.46% (20954/26371) |
| Line Coverage     | 72.54% (216510/298489) |
| Region Coverage   | 70.78% (127448/180057) |
| Branch Coverage   | 64.48% (65997/102350) |