@hello-stephen 2025-06-04
# BE Regression && UT Coverage Report
Increment line coverage `100.00% (3/3)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_f13b28b70f73573afab942cb7c38cab126b7606d_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_f13b28b70f73573afab942cb7c38cab126b7606d_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 79.49% (20962/26372) |
| Line Coverage     | 72.54% (216519/298492) |
| Region Coverage   | 70.74% (127370/180058) |
| Branch Coverage   | 64.43% (65942/102350) |