@hello-stephen 2025-06-05
# FE UT Coverage Report
Increment line coverage `  73.54% (1779/2419)` :tada:
[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_ab44b95e0d82742ba66f88e53a6832b316136a48/fe_increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_ab44b95e0d82742ba66f88e53a6832b316136a48/fe_report/index.html)