@hello-stephen 2025-06-06
# FE UT Coverage Report
Increment line coverage `  73.49% (1780/2422)` :tada:
[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_61e7e1014e6c0e8fed3edac4a6b0f4ea90fb883e/fe_increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_61e7e1014e6c0e8fed3edac4a6b0f4ea90fb883e/fe_report/index.html)