@hello-stephen 2025-06-13
# FE UT Coverage Report
Increment line coverage `  73.17% (1789/2445)` :tada:
[Increment coverage report](http://coverage.selectdb-in.cc/coverage/43546_7f15e1dd74e81109925fef4a85ffeb76425bd63b/fe_increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/43546_7f15e1dd74e81109925fef4a85ffeb76425bd63b/fe_report/index.html)